Move the functions for reading room state to a separate package.
This should:
    1) Make the input package a bit cleaner.
    2) Allow use to reuse the state reading code from the query package.
